I added a bunch of prints everywhere
To see them open the serial monitor under tools. There's an infinite loop inside the transaction for loop for reading data, if you comment it out you will get a nasty error, for some strange reason the data read is causing it (im 90% sure its the data read).  Everything before that is working, but when it comes to that, it goes all wrong.